==> script01_select_behav_data.ipynb <==
import pandas as pd
import numpy as np
    
# reading behavioral scores:
subject_behav = pd.read_csv("data/pnc_cognitive_response.csv", delimiter = ",")


#print(subject_info_behav[subject_info_behav.duplicated(subset = "SUBJID", keep = False)])

cog_label_acc = [
    # executive control:
    "pcet_acc2",
    "pcpt_t_tp",
    "lnb_tp2",
    # episodic memory task:
    "pwmt_kiwrd_tot",
    "pfmt_ifac_tot",
    "volt_svt",
    # complex cognitive task:
    "pvrt_cr",
    "pmat_cr",
    "plot_tc",
    # social cognition task:
    "peit_cr",
    "pedt_a",
    "padt_a"
]

cog_label_speed = [
    # executive control:
    "pcet_rtcr",
    "pcpt_t_tprt",
    "lnb_rtc2",
    # episodic memory task:
    "pwmt_kiwrd_rtc",
    "pfmt_ifac_rtc",
    "volt_svtcrt",
    # complex cognitive task:
    "pvrt_rtcr",
    "pmat_rtcr",
    "plot_tcrt",
    # social cognition task:
    "peit_crt",
    "pedt_t",
    "padt_t",
    # sensorimotor speed task:
    "mp_mp2rtcr",
    "tap_tot"
]

cog_label_acc2 = [
    # executive control:
    "ABF",
    "ATT",
    "WM",
    # episodic memory task:
    "VMEM",
    "FMEM",
    "SMEM",
    # complex cognitive task:
    "LAN",
    "NVR",
    "SPA",
    # social cognition task:
    "EID",
    "EDI",
    "ADI"
]

cog_label_speed2 = [
    # executive control:
    "ABF",
    "ATT",
    "WM",
    # episodic memory task:
    "VMEM",
    "FMEM",
    "SMEM",
    # complex cognitive task:
    "LAN",
    "NVR",
    "SPA",
    # social cognition task:
    "EID",
    "EDI",
    "ADI",
    # sensorimotor speed task:
    "MOT",
    "SM"
]

cog_label = cog_label_acc + cog_label_speed
cog_label2 = [s + '_Accuracy' for s in cog_label_acc2] + [s + '_Speed' for s in cog_label_speed2]

#print(subject_info_behav[["mp_mp2rtcr", "tap_tot"]])
# revert the speed cognitive scores:
subject_behav[cog_label_speed] = subject_behav[cog_label_speed]*(-1)



subject_behav_selected = subject_behav.loc[:, ['SUBJID', 'Sex']+cog_label]
subject_behav_selected.columns = ['SUBJID', 'Sex']+cog_label2
subject_behav_selected.to_csv('result/out01_behav_selected.csv', sep = ',')
