==> script03_regression_rfe_disorders_show_results.ipynb <==
# run script02_age_prediction_disorders for all four methods and show regression rfe results in one file:
import pandas as pd

# regression_rfe_table_dnn = pd.read_csv("result/disorders/regression_rfe_DNN_GMV&DTI.csv")
# regression_rfe_table_ridge = pd.read_csv("result/disorders/regression_rfe_Ridge_GMV&DTI.csv")
# regression_rfe_table_gpr = pd.read_csv("result/disorders/regression_rfe_GPR_GMV&DTI.csv")
# regression_rfe_table_svr = pd.read_csv("result/disorders/regression_rfe_SVR_GMV&DTI.csv") 

regression_rfe_table_dnn = pd.read_csv("result/disorders/regression_rfe_DNN_Multi-modal.csv")
regression_rfe_table_ridge = pd.read_csv("result/disorders/regression_rfe_Ridge_Multi-modal.csv")
regression_rfe_table_gpr = pd.read_csv("result/disorders/regression_rfe_GPR_Multi-modal.csv")
regression_rfe_table_svr = pd.read_csv("result/disorders/regression_rfe_SVR_Multi-modal.csv") 


regression_rfe_table_dnn=regression_rfe_table_dnn.reindex([2, 0, 1, 3, 4, 5])
regression_rfe_table_dnn

regression_rfe_table_ridge

regression_rfe_table_gpr

regression_rfe_table_svr
